feat: add AdminRoute component, edit ProtectedRoute component, refactor adminDataSlice
