+++ src/components/AdminRoute/AdminRoute.jsx
@@ -2,10 +2,15 @@ import { Navigate, Outlet } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { ENDPOINT_ROUTES } from '../../constants/constantsEndpointRoute.js';
 
-function ProtectedRoute() {
+function AdminRoute() {
   const isLoggedIn = useSelector((state) => state.isLoggedIn.isLoggedIn);
+  const role = useSelector((state) => state.adminData.role);
 
-  return isLoggedIn ? <Outlet /> : <Navigate to={ENDPOINT_ROUTES.login} />;
+  return isLoggedIn && role === 'ROLE_ADMIN' ? (
+    <Outlet />
+  ) : (
+    <Navigate to={ENDPOINT_ROUTES.login} />
+  );
 }
 
-export default ProtectedRoute;
+export default AdminRoute;
