+++ src/components/SideMenu/SideMenu.jsx
@@ -5,14 +5,14 @@ import MenuEmployee from './MenuEmployee/MenuEmployee.jsx';
 import exitIcon from '../../images/exit_button.svg';
 import logo from '../../images/logo.svg';
 import { setIsLoggedIn } from '../../store/slices/isLoggedInSlice.js';
-import { setAdminData } from '../../store/slices/adminDataSlices.js';
+import { setAdminData } from '../../store/slices/adminDataSlice.js';
 import { setToken } from '../../store/slices/tokenSlices.js';
 import './SideMenu.scss';
 
 function SideMenu() {
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const adminData = useSelector((state) => state.adminData.adminData);
+  const adminData = useSelector((state) => state.adminData);
 
   const handleLogout = (e) => {
     e.preventDefault();
